+++ packages/tswebextension/src/lib/common/index.ts
@@ -7,3 +7,4 @@ export * from './utils';
 export * from './stealth-helper';
 export * from './content-script/send-app-message';
 export * from './request-type';
+export * from './error';