+++ packages/tswebextension/test/lib/mv2/background/fixtures/local-script-rules.ts
@@ -1,4 +1,6 @@
-export default {
+import { LocalScriptRules } from '@lib/mv2/background/services/local-script-rules-service';
+
+export const getLocalScriptRulesFixture = (): LocalScriptRules => ({
     comment: 'Test list of pre-built JS rules',
     rules: [
         {
@@ -6,4 +8,4 @@ export default {
             script: 'window.open = undefined;',
         },
     ],
-};
+});